+++ src/app/member/component.spec.ts
@@ -32,7 +32,7 @@ import {SettingsService} from '@core/services/settings/service';
 import {UserService} from '@core/services/user/service';
 import {SharedModule} from '@shared/shared.module';
 import {of} from 'rxjs';
-import {MemberComponent} from './member.component';
+import {MemberComponent} from './component';
 
 describe('MemberComponent', () => {
   let fixture: ComponentFixture<MemberComponent>;
